events improve

+++ src/main/java/com/mmm/podobri/dao/BaseDao.java
@@ -11,12 +11,12 @@ import java.util.List;
 
 public interface BaseDao<T>
 {
-    public T findOne(final long id);
+    public T findOne(final int id);
     public List<T> findAll();
     public void save(final T entity);
     public T update(final T entity);
     public void delete(final T entity);
-    public void deleteById(final long id);
+    public void deleteById(final int id);
     public DaoUtils getDaoUtils();
     public void saveInTransaction(Serializable... entities);
 }
