+++ src/main/java/com/mmm/podobri/service/BaseDaoServices.java
@@ -13,12 +13,12 @@ import com.mmm.podobri.dao.DaoUtils;
 
 public interface BaseDaoServices<T>
 {
-    public T findOne(final long id);
+    public T findOne(final int id);
     public List<T> findAll();
     public void save(final T entity);
     public T update(final T entity);
     public void delete(final T entity);
-    public void deleteById(final long id);
+    public void deleteById(final int id);
     public DaoUtils getDaoUtils();
     public void saveInTransaction(Serializable... entities);
 }
